Promove main (5f0f4240) para produção
Regenerado por scripts/promote-producao.sh: vendor/ sem dev, public/build/
compilado, sem tests/ nem node_modules/, .gitignore próprio.

+++ app/Http/Middleware/IdentifyTenant.php
@@ -30,7 +30,7 @@ class IdentifyTenant
         $tenant = config('tenancy.cache.enabled')
             ? Cache::remember(
                 "tenant:slug:{$slug}",
-                now()->addMinutes(config('tenancy.cache.ttl_minutes')),
+                now()->addMinutes((int) config('tenancy.cache.ttl_minutes')),
                 $resolveTenant,
             )
             : $resolveTenant();
